@@ -2,29 +2,28 @@
   "name": "moncomptepro",
   "version": "1.0.0",
   "main": "src/index.js",
+  "type": "module",
   "engines": {
-    "node": "20.11.1"
+    "node": "20.12.2"
   },
   "scripts": {
     "lint": "prettier --write */*.ts",
-    "start": "node build/index.js",
+    "start": "tsx src/index.ts",
     "dev": "NODE_ENV=development concurrently \"npm:watch:*\"",
     "build": "run-s build:**",
-    "build:typescript": "tsc",
-    "build:views": "bash -c 'cp -r src/views build/views'",
     "build:assets": "vite build --clearScreen false",
     "build:db:migrate": "node-pg-migrate up",
     "build:db:conditionaly-load-fixtures": "if [ \"$RUN_FIXTURES\" = \"True\" ]; then npm run fixtures:load; fi",
-    "update-organization-info": "ts-node --files scripts/update-organizations-info.ts",
+    "update-organization-info": "tsx scripts/update-organizations-info.ts",
     "migrate": "node-pg-migrate",
-    "fixtures:load": "node-pg-migrate up && ts-node scripts/run-sql-script.ts scripts/fixtures.sql && npm run update-organization-info 2000",
-    "fixtures:load-ci": "node-pg-migrate up && ts-node scripts/run-sql-script.ts",
-    "delete-database": "if [ \"$ENABLE_DATABASE_DELETION\" = \"True\" ]; then ts-node scripts/run-sql-script.ts scripts/delete_all.sql; fi",
+    "fixtures:load": "node-pg-migrate up && tsx scripts/run-sql-script.ts scripts/fixtures.sql && npm run update-organization-info 2000",
+    "fixtures:load-ci": "node-pg-migrate up && tsx scripts/run-sql-script.ts",
+    "delete-database": "if [ \"$ENABLE_DATABASE_DELETION\" = \"True\" ]; then tsx scripts/run-sql-script.ts scripts/delete_all.sql; fi",
     "test": "run-s test:*",
     "test:lint": "prettier */*.ts --list-different",
-    "test:unit": "CONSIDER_ALL_EMAIL_DOMAINS_AS_NON_FREE=False DO_NOT_USE_ANNUAIRE_EMAILS=False DO_NOT_CHECK_EMAIL_DELIVERABILITY=False ts-mocha -p tsconfig.json test/*.ts",
+    "test:unit": "CONSIDER_ALL_EMAIL_DOMAINS_AS_NON_FREE=False DO_NOT_USE_ANNUAIRE_EMAILS=False DO_NOT_CHECK_EMAIL_DELIVERABILITY=False mocha --require tsx test/*.ts",
     "watch:js": "tsc --watch --preserveWatchOutput",
-    "watch:node": "nodemon --watch build --ext js,ejs --quiet build/index.js",
+    "watch:node": "tsx --watch src/index.ts",
     "watch:ejs": "CHOKIDAR_USEPOLLING=true copy-and-watch --clean --watch \"src/views/**/*.ejs\" src/views/*.ejs build/views/",
     "watch:assets": "NODE_ENV=development vite build --clearScreen false --watch"
   },
